+++ Cleaner/MbScrubCleaner.php
@@ -1,0 +1,37 @@
+<?php
+
+/**
+ * Part of EncodingRepair package.
+ *
+ * (c) Adrien Loyant <donald_duck@team-df.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Ducks\Component\EncodingRepair\Cleaner;
+
+/**
+ * Cleaner using mb_scrub to remove invalid sequences.
+ *
+ * @psalm-api
+ */
+final class MbScrubCleaner implements CleanerInterface
+{
+    public function clean(string $data, string $encoding, array $options): ?string
+    {
+        return \mb_scrub($data, $encoding);
+    }
+
+    public function getPriority(): int
+    {
+        return 100;
+    }
+
+    public function isAvailable(): bool
+    {
+        return \function_exists('mb_scrub');
+    }
+}